requirements.txt
# Django and REST Framework
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1

# Environment variables
python-dotenv==1.0.0

# Database (optional - only if using PostgreSQL)
# psycopg2-binary==2.9.9  # Uncomment if you need PostgreSQL support

# Google AI/ADK dependencies
google-adk==0.1.0  # Adjust version as needed
google-generativeai==0.3.2

# Async support
asgiref==3.7.2

# Additional utilities
python-dateutil==2.8.2
pytz==2024.1

# Development tools (optional)
black==24.1.1
flake8==7.0.0
pylint==3.0.3
